Remove Cupertino View and BottomBar
